掲示板システム制作課題
削除済みの投稿・コメントに対する操作を行った場合の処理を実装していなかった
・操作対象の投稿・コメントが削除されていて存在しない場合の処理を追加

NGワードの追加機能で、既にNGワードに追加されているフレーズを含んだ単語を追加しようとするとエラーになる
・NGワードの重複チェックの処理を修正(IndexOfメソッドからequalsメソッドに変更)

新規投稿時にカテゴリが空白で埋まっている状態でエラーにならない
・バリデーションチェックの項目に追加

ユーザ情報変更時のログインID重複チェックが不十分だった
・コードを修正

+++ src/bbs/controller/DeleteCommentServlet.java
@@ -1,6 +1,7 @@
 package bbs.controller;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.servlet.ServletException;
@@ -17,6 +18,11 @@ import bbs.service.MessageService;
 @WebServlet(urlPatterns = {"/deleteComment"})
 public class DeleteCommentServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
+
+	@Override
+	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException{
+		response.sendRedirect("./top");
+	}
 	@Override
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException{
 
@@ -24,10 +30,15 @@ public class DeleteCommentServlet extends HttpServlet {
 		int id = Integer.parseInt(request.getParameter("id"));
 
 		HttpSession session = request.getSession();
+		List<String> errorMessages = new ArrayList<String>();
 
+		if(isExistComment(id)){
+			new MessageService().deleteComment(id);
+		}else{
+			errorMessages.add("削除しようとしたコメントは存在しません");
+			session.setAttribute("errorMessages", errorMessages);
+		}
 
-
-		new MessageService().deleteComment(id);
 
 
 		List<String> categories = new MessageService().getCategories();
@@ -37,7 +48,10 @@ public class DeleteCommentServlet extends HttpServlet {
 		session.setAttribute("categories", categories);
 		List<Comment> comments = new MessageService().getComment();
 		session.setAttribute("comments", comments);
+		request.setAttribute("postCount", messages.size());
 		response.sendRedirect("./top");
 	}
-
+	private boolean isExistComment(int id) {
+		return new MessageService().isExistComment(id);
+	}
 }
